requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.5
django-crispy-forms==2.3
dnspython==2.7.0
ldap3==2.9.1
ms_active_directory==1.14.1
packaging==24.2
pip-review==1.3.0
pyasn1==0.6.1
pycryptodome==3.21.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
sqlparse==0.5.3
